533 - add delete button

+++ src/Application/Command/Type/RemoveHandler.php
@@ -10,6 +10,7 @@
 
 namespace Proximum\Vimeet\Application\Command\Type;
 
+use Proximum\Vimeet\Domain\Repository\SheetRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\TypeRepositoryInterface;
 
 class RemoveHandler
@@ -20,17 +21,35 @@ class RemoveHandler
     private $typeRepository;
 
     /**
+     * @var SheetRepositoryInterface
+     */
+    private $sheetRepository;
+
+    /**
      * RemoveHandler constructor.
      *
-     * @param TypeRepositoryInterface $typeRepository
+     * @param TypeRepositoryInterface  $typeRepository
+     * @param SheetRepositoryInterface $sheetRepository
      */
-    public function __construct(TypeRepositoryInterface $typeRepository)
-    {
-        $this->typeRepository = $typeRepository;
+    public function __construct(
+        TypeRepositoryInterface $typeRepository,
+        SheetRepositoryInterface $sheetRepository
+    ) {
+        $this->typeRepository  = $typeRepository;
+        $this->sheetRepository = $sheetRepository;
     }
 
+    /**
+     * @param Remove $remove
+     *
+     * @return null|bool
+     */
     public function handle(Remove $remove)
     {
+        if ($this->sheetRepository->isThereAtLeastOneByType($remove->type)) {
+            return false;
+        }
+
         $this->typeRepository->remove($remove->type);
     }
 }
